Renames slopecontrollers to speedmodifiers.
