--- subsystems/motor/slopecontrollers/SlopeController.java
@@ -1,6 +1,0 @@
-package org.usfirst.frc4904.standard.subsystems.motor.slopecontrollers;
-
-
-public interface SlopeController {
-	double reslope(double speed);
-}
