+++ subsystems/motor/speedmodifiers/AccelerationCap.java
@@ -1,9 +1,9 @@
-package org.usfirst.frc4904.standard.subsystems.motor.slopecontrollers;
+package org.usfirst.frc4904.standard.subsystems.motor.speedmodifiers;
 
 
 import org.usfirst.frc4904.standard.custom.sensors.PDP;
 
-public class AccelerationCap implements SlopeController {
+public class AccelerationCap implements SpeedModifier {
 	private double currentSpeed;
 	private long lastUpdate;
 	private final PDP pdp;
@@ -20,7 +20,7 @@ public class AccelerationCap implements SlopeController {
 		this(pdp, 11.0, 10.0);
 	}
 	
-	public double reslope(double speed) {
+	public double modify(double speed) {
 		if (Math.abs(speed) > Math.abs(currentSpeed) && pdp.getVoltage() < softStopVoltage) {
 			speed = currentSpeed;
 			if (pdp.getVoltage() < hardStopVoltage) {
