+++ subsystems/motor/speedmodifiers/SpeedModifier.java
@@ -1,0 +1,6 @@
+package org.usfirst.frc4904.standard.subsystems.motor.speedmodifiers;
+
+
+public interface SpeedModifier {
+	double modify(double speed);
+}
